# vmagent: improve streamParse perofrmance