### Describe Your Changes

Enhance concurrent stream parsing performance in vmagent by:
- Removing global lock around stream.Parse callback.
- Using atomic operations for counters
- Creating write contexts per callback instead of sharing
- Improving series limit checking with sync.Once
- Optimizing labels hash calculation with buffer pooling
- Adding comprehensive tests for concurrency correctness

Fixes https://github.com/VictoriaMetrics/VictoriaMetrics/issues/8159

```
$benchstat master.bench pr-8515.bench 
goos: darwin
goarch: arm64
pkg: github.com/VictoriaMetrics/VictoriaMetrics/lib/promscrape
cpu: Apple M1 Pro
                                         │ master.bench │            pr-8515.bench            │
                                         │    sec/op    │   sec/op     vs base                │
GetScrapeWork-10                           641.5n ± 14%   560.3n ± 8%  -12.66% (p=0.007 n=10)
IsAutoMetricMiss-10                        2.579n ±  5%   2.001n ± 3%  -22.43% (p=0.000 n=10)
IsAutoMetricHit-10                         4.244n ± 14%   3.385n ± 3%  -20.23% (p=0.000 n=10)
ScrapeWorkScrapeInternalOneShot-10         1.227µ ± 12%   1.155µ ± 7%   -5.91% (p=0.003 n=10)
ScrapeWorkScrapeInternalStream-10          3.924µ ±  3%   3.516µ ± 5%  -10.40% (p=0.000 n=10)
ScrapeWorkScrapeInternalStreamBigData-10   90.51m ±  1%   16.72m ± 2%  -81.53% (p=0.000 n=10)
ScrapeWorkGetLabelsHash-10                 585.0n ±  1%   566.7n ± 2%   -3.15% (p=0.002 n=10)
geomean                                    1.087µ         759.6n       -30.10%

                                         │  master.bench  │             pr-8515.bench              │
                                         │      B/s       │      B/s       vs base                 │
GetScrapeWork-10                            1.488Mi ± 12%    1.702Mi ± 9%   +14.42% (p=0.006 n=10)
IsAutoMetricMiss-10                         369.7Mi ±  5%    476.7Mi ± 3%   +28.95% (p=0.000 n=10)
IsAutoMetricHit-10                          224.7Mi ± 13%    281.7Mi ± 3%   +25.35% (p=0.000 n=10)
ScrapeWorkScrapeInternalOneShot-10         1015.0Mi ± 11%   1078.6Mi ± 6%    +6.27% (p=0.003 n=10)
ScrapeWorkScrapeInternalStream-10           317.4Mi ±  3%    354.3Mi ± 5%   +11.62% (p=0.000 n=10)
ScrapeWorkScrapeInternalStreamBigData-10    32.55Mi ±  1%   176.22Mi ± 2%  +441.42% (p=0.000 n=10)
geomean                                     104.4Mi          157.7Mi        +51.05%

                                         │  master.bench  │             pr-8515.bench              │
                                         │      B/op      │     B/op      vs base                  │
GetScrapeWork-10                             865.0 ± 0%       865.0 ± 0%        ~ (p=1.000 n=10)
IsAutoMetricMiss-10                          0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
IsAutoMetricHit-10                           0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
ScrapeWorkScrapeInternalOneShot-10           0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
ScrapeWorkScrapeInternalStream-10          9.046Ki ± 0%     9.013Ki ± 0%   -0.36% (p=0.000 n=10)
ScrapeWorkScrapeInternalStreamBigData-10   18.15Mi ± 2%     32.96Mi ± 1%  +81.61% (p=0.000 n=10)
ScrapeWorkGetLabelsHash-10                   0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
geomean                                                 ²                  +8.84%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                         │ master.bench │             pr-8515.bench             │
                                         │  allocs/op   │ allocs/op   vs base                   │
GetScrapeWork-10                           6.000 ± 0%     6.000 ± 0%         ~ (p=1.000 n=10) ¹
IsAutoMetricMiss-10                        0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
IsAutoMetricHit-10                         0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
ScrapeWorkScrapeInternalOneShot-10         0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
ScrapeWorkScrapeInternalStream-10          25.00 ± 0%     25.00 ± 0%         ~ (p=1.000 n=10) ¹
ScrapeWorkScrapeInternalStreamBigData-10   186.5 ± 3%     633.5 ± 1%  +239.68% (p=0.000 n=10)
ScrapeWorkGetLabelsHash-10                 0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=10) ¹
geomean                                               ²                +19.09%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

### Checklist

The following checks are **mandatory**:

- [x] My change adheres [VictoriaMetrics contributing guidelines](https://docs.victoriametrics.com/contributing/).