@hagen1778 2025-03-20
> @hagen1778 should I rebase&squash?

We can solve minor conflicts, as one in the changelog, on our own. And we squash on merge.