@makasim 2025-03-20
> You can squash commits, sign result commit and force-push it back to this pull request.

@f41gh7 done, it shows `partially signed` mark near the comment, is this okay ?