@hagen1778 2025-03-20
Looks like one of the tests is flaky:
```
--- FAIL: TestScrapeWorkScrapeInternalStreamConcurrency (1.96s)
    scrapework_test.go:612: unexpected number of pushed timeseries; got 4014; want 4015
```
https://github.com/VictoriaMetrics/VictoriaMetrics/actions/runs/13973594574/job/39122102833#step:5:89